requirements.txt
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dataclass==0.0
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.5.0
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
toml==0.10.0
typing==3.7.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
